Updated read only color interface to return the alpha value as an int

+++ src/main/java/model/color/IReadOnlyColor.java
@@ -27,11 +27,11 @@ public interface IReadOnlyColor {
   int getBlue();
 
   /**
-   * Returns the alpha value in the range [0, 1].
+   * Returns the alpha value in the range [0, 255].
    *
    * @return the alpha value.
    */
-  double getAlpha();
+  int getAlpha();
 
   /**
    * Returns the red percentage of the color in the range [0, 1].
